Resolve issue #224.
Force a poll of the project file write times before any resource gathering is being done from the solution
explorer. This ensures that Ankh's view of the state of project files is correct when invoking an SVN operation, 
and ensures modified project files are included in a commit, even if the user didn't explicitly save them first.

* \src\Ankh\Solution\Explorer.cs
	(Refresh, Get*Resources, VisitSelectedNodes) - Insert calls to ForcePoll()
	(ForcePoll) - New method, invokes the new ForcePoll  method on the ProjectFileWatcher.

* \src\Ankh\FileWatcher.cs
	(ForcePoll) - New method, forces a poll.