Integrated the two codepaths that update the statuscache to one.
The new private RefreshPath() now receives all updates. 
This function guaranties that the information on the specified path value 
is updated allowing a query of the specified depth.

The statuscall uses the specified path (or its parent folder if it is a file and
the depth is greater than Empty)

If depth is at least files the deleted files in the folder are updated (if depth
is children or indefinate deleted folders are updated to)

When the statuscall turns up empty (no results on the specified path) a non svn 
crawl is used on the specified folder (like the old refresh).

SvnItem contains 5 individual caches now: ankhstatus, svn-mustlock, OnDisk-IsFile, OnDisk-Exists and OnDisk-Readonly.
MarkDirty() marks them all dirty, but their status can be changed independent.
(The last three are now resolved by a direct GetFileAttributesW() call to bypass several layers of security normalization inside .Net)

* Ankh/StatusCache/FileStatusCache.cs
* Ankh/StatusCache/StatusCache.cs
  Integrated 2.5 statuswalkers into 1; optimizing the SharpSvn experience
  by using a callback which updates as much of the state as we possibly can.

* Ankh.Services/Scc/AnkhStatus.cs
  Replaced some if's by switched (which are optimized by the c# compiler)
  Added DebuggerDisplay attribute  

* Ankh.Scc/AnkhSccProvider.Glyphs.cs
  Use SvnItem.IsConflicted

* Ankh.Services/Ankh.Services.csproj
  Added IWorkingCopyOperations and upgraded SharpSvn reference

* Ankh.Services/Scc/IFileStatusCache.cs
  Renamed RefreshMe() to RefreshItem() and applied more comments

* Ankh.Services/Scc/IWorkingCopyOperations.cs
  Moved old service here

* Ankh.Services/Scc/SvnItem.cs
  Use an XBool try state value for caching exists, readonly, mustlock and isFile outside
  the AnkhStatus. Refresh them via GetFileAttributes() to shortcut .Net security parsing.

* Ankh.Tests/ContextBase.cs
  Removed unused reference to status cache

* Ankh/Ankh.csproj
* Ankh/StatusCache/DeletedFilesList.cs
  Added KeyedCollection for deleted files

* */* 
  Namespace renames for IWorkingCopyOperations and property renames
  on SvnItem